Tyuiu.AvdeevAS.Sprint5.Task2.V7.Lib/DataService.cs
using tyuiu.cources.programming.interfaces.Sprint5;
namespace Tyuiu.AvdeevAS.Sprint5.Task2.V7.Lib
{
    public class DataService : ISprint5Task2V7
    {
        public string SaveToFileTextData(int[,] matrix)
        {
            string path = Path.Combine(Path.GetTempPath(), "OutPutFileTask2.csv");

            for (int i = 0; i < matrix.GetLength(0); i++)
            {
                for (int j = 0; j < matrix.GetLength(1); j++)
                {
                    // Заменяем нечетные элементы на 0
                    if (matrix[i, j] % 2 != 0)
                    {
                        matrix[i, j] = 0;
                    }
                }
            }

            using (StreamWriter writer = new StreamWriter(path))
            {
                for (int i = 0; i < matrix.GetLength(0); i++)
                {
                    for (int j = 0; j < matrix.GetLength(1); j++)
                    {
                        writer.Write(matrix[i, j]);
                        if (j < matrix.GetLength(1) - 1)
                        {
                            writer.Write(";"); // Разделитель для CSV
                        }
                    }
                    writer.WriteLine(); // Переход на новую строку
                }
            }

            return path;
        }
    }
}

